chore(mockgen): move model to internal
The mockgen/model package was importable only for reflect mode.
In reflect mode, the generated code would import model,
which would then use reflection to export information about types.

Reflect mode is not a thing anymore,
so there's no need for model to be externally accessible.

Move it into an internal package and update all imports
*except* those inside the bazel/ directory for the following reasons:

- bazel/ is pinned to an older version of this package
- those references will be deleted in #285
