+++ mockgen/generic.go
@@ -13,7 +13,7 @@ import (
 	"go/ast"
 	"go/token"
 
-	"go.uber.org/mock/mockgen/model"
+	"go.uber.org/mock/mockgen/internal/model"
 )
 
 func getTypeSpecTypeParams(ts *ast.TypeSpec) []*ast.Field {
